[gw] GWSolver with tests
